Fixed data-repeat bug and implemented chunked loading
Stupid, STUPID bug. The `file.seek` method requires a number of bytes,
not of samples. DUH! In addition, data is now loaded in chunks of 1000
samples, which should keep the amount of memory used at a time in check.
(Whether 32-bit systems overload now depends on whether the `data`
that's passed to `_read_segment_file` is a memory-mapped array. If it
isn't, there will still be a point where Python will run out of memory.)